[feat] 동행글 좋아요 등록 (#28)
동행글 좋아요 기능 추가

동행글 좋아요 등록 조회 C, S, R 구현 완료
동행글 좋아요 등록 조회 관련 DTO 구현 완료

+++ src/main/java/com/example/ballkkaye/match/chat/room/user/ChatRoomUserRepository.java
@@ -32,12 +32,12 @@ public class ChatRoomUserRepository {
         }
     }
 
-    public Long countByChatRoomId(Integer chatRoomId) {
+    public Integer countByChatRoomId(Integer chatRoomId) {
         return em.createQuery("""
                         SELECT COUNT(cru) FROM ChatRoomUser cru
                         WHERE cru.chatRoom.id = :chatRoomId AND cru.deleteStatus = 'NOT_DELETED'
                         """, Long.class)
                 .setParameter("chatRoomId", chatRoomId)
-                .getSingleResult();
+                .getSingleResult().intValue();
     }
 }
